move request_id to utils.go
